style: 충돌방지, theme
feature/report 브랜치의 현재 theme를 가져오고 StyledTable끼리 충돌 안나게 함.
