+++ components/styledComponents/StyledTableUser1.tsx
@@ -10,13 +10,19 @@ import TableHead from "@mui/material/TableHead";
 import TablePagination from "@mui/material/TablePagination";
 import TableRow from "@mui/material/TableRow";
 
-export default function StickyHeadTable3({
+export default function StickyHeadTableUser1({
   columns,
   rows,
 }: {
   columns: string[];
   rows: { [key: string]: string | number }[];
 }) {
+  const router = useRouter();
+
+  const handleRowClick = (userId: string | number) => {
+    router.push(`/user/${userId}`);
+  };
+
   return (
     <Paper sx={{ width: "100%", overflow: "hidden" }}>
       <TableContainer sx={{ maxHeight: 800 }}>
@@ -33,18 +39,20 @@ export default function StickyHeadTable3({
           <TableBody>
             {rows.length === 0 ? (
               <TableRow>
-                <TableCell
-                  colSpan={columns.length}
-                  align="center"
-                  sx={{ height: "200px" }}
-                >
+                <TableCell colSpan={columns.length} align="center">
                   검색 결과가 없습니다.
                 </TableCell>
               </TableRow>
             ) : (
               rows.map((row, rowIndex) => {
                 return (
-                  <TableRow hover role="checkbox" tabIndex={-1} key={rowIndex}>
+                  <TableRow
+                    hover
+                    role="checkbox"
+                    tabIndex={-1}
+                    key={rowIndex}
+                    onClick={() => handleRowClick(row["닉네임"])}
+                  >
                     {columns.map((column, index) => {
                       const value = row[column];
                       return (
