+++ components/styledComponents/StyledTableUser2.tsx
@@ -10,19 +10,13 @@ import TableHead from "@mui/material/TableHead";
 import TablePagination from "@mui/material/TablePagination";
 import TableRow from "@mui/material/TableRow";
 
-export default function StickyHeadTable2({
+export default function StickyHeadTableUser2({
   columns,
   rows,
 }: {
   columns: string[];
   rows: { [key: string]: string | number }[];
 }) {
-  const router = useRouter();
-
-  const handleRowClick = (userId: string | number) => {
-    router.push(`/user/${userId}`);
-  };
-
   return (
     <Paper sx={{ width: "100%", overflow: "hidden" }}>
       <TableContainer sx={{ maxHeight: 800 }}>
@@ -39,20 +33,18 @@ export default function StickyHeadTable2({
           <TableBody>
             {rows.length === 0 ? (
               <TableRow>
-                <TableCell colSpan={columns.length} align="center">
+                <TableCell
+                  colSpan={columns.length}
+                  align="center"
+                  sx={{ height: "200px" }}
+                >
                   검색 결과가 없습니다.
                 </TableCell>
               </TableRow>
             ) : (
               rows.map((row, rowIndex) => {
                 return (
-                  <TableRow
-                    hover
-                    role="checkbox"
-                    tabIndex={-1}
-                    key={rowIndex}
-                    onClick={() => handleRowClick(row["닉네임"])}
-                  >
+                  <TableRow hover role="checkbox" tabIndex={-1} key={rowIndex}>
                     {columns.map((column, index) => {
                       const value = row[column];
                       return (
